+++ cpp/src/hub/solver/py_bfws.cc
@@ -54,7 +54,7 @@ public :
                         throw std::invalid_argument("AIRLAPS exception: BFWS algorithm needs python states for implementing __hash__()");
                     }
                     // python __hash__ can return negative integers but c++ expects positive integers only
-                    return s._state.attr("__hash__")().template cast<long>() + std::numeric_limits<long>::max();
+                    return s._state.attr("__hash__")().template cast<long long>() + std::numeric_limits<long long>::max();
                 } catch(const py::error_already_set& e) {
                     throw std::runtime_error(e.what());
                 }
@@ -103,7 +103,7 @@ public :
                         throw std::invalid_argument("AIRLAPS exception: BFWS algorithm needs python events for implementing __hash__()");
                     }
                     // python __hash__ can return negative integers but c++ expects positive integers only
-                    return e._event.attr("__hash__")().template cast<long>() + std::numeric_limits<long>::max();
+                    return e._event.attr("__hash__")().template cast<long long>() + std::numeric_limits<long long>::max();
                 } catch(const py::error_already_set& ex) {
                     throw std::runtime_error(ex.what());
                 }
